Extracting the Segment::Fields from the Segment class
