+++ lib/ruby-hl7.rb
@@ -52,7 +52,9 @@ require 'message_parser'
 require 'message'
 require 'segment_list_storage'
 require 'segment_generator'
+require 'segment_fields'
 require 'segment'
+require 'segment_default'
 
 require 'core_ext/date_time'
 require 'core_ext/string'
